test: split cli test file to match code organization
Move TestGenerateCerts to generate_certs_test.go and TestRunUncors
to run_uncors_test.go. Tests are now co-located with the functions
they test, matching the split of commands.go into separate modules.

Also move helper functions (httpMapping, waitForPort, writeConfig,
startProxy) to run_uncors_test.go where they are used.
