+++ internal/cli/generate_certs_test.go
@@ -1,0 +1,23 @@
+package cli_test
+
+import (
+	"testing"
+
+	"github.com/evg4b/uncors/internal/cli"
+	"github.com/stretchr/testify/require"
+)
+
+func TestGenerateCerts(t *testing.T) {
+	t.Run("returns error for unknown flag", func(t *testing.T) {
+		err := cli.GenerateCerts([]string{"--unknown-flag"})
+		require.Error(t, err)
+	})
+
+	t.Run("generates CA certificate with valid args", func(t *testing.T) {
+		// Point HOME to a temp dir so certs go there, not ~/.config/uncors.
+		t.Setenv("HOME", t.TempDir())
+
+		err := cli.GenerateCerts([]string{"--validity-days=7"})
+		require.NoError(t, err)
+	})
+}
